moved tests to matching package structure
