+++ src/test/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/TurtleGeneratorTest.java
@@ -1,15 +1,13 @@
-package com.marginallyclever.makelangeloRobot;
+package com.marginallyclever.makelangelo.makeart.turtlegenerator;
 
 import com.marginallyclever.makelangelo.Translator;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGenerator;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGeneratorFactory;
 import com.marginallyclever.util.PreferencesHelper;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class MakelangeloRobotPanelTest {
+public class TurtleGeneratorTest {
 
     @Test
     public void testNoMissingGeneratorPanels() {
@@ -17,7 +15,7 @@ public class MakelangeloRobotPanelTest {
         Translator.start();
         try {
             for (TurtleGenerator c : TurtleGeneratorFactory.available) {
-                assertNotNull(c.getPanel());
+                assertNotNull(new TurtleGeneratorPanel(c));
             }
         } catch (Exception e) {
             fail("Missing panel! " + e.getLocalizedMessage());
